requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
requests==2.31.0
gunicorn==21.2.0